Refactoring

+++ include/space.hpp
@@ -9,30 +9,30 @@
 #include <cmath>
 #include <functional>
 
-namespace linalg{
+namespace linalg {
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T product(const Vector<N, T>& a, const Vector<N, T>& b);
 
-    template <size_t P, size_t N, typename T>
+    template <int P, int N, typename T>
     T norm(const Vector<N, T>& vector);
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T euclidian(const Vector<N, T>& vector);
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T manhattan(const Vector<N, T>& vector);
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     using NormAlgorithm = T(*)(const linalg::Vector<N, T>&);
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T distance(Vector<N, T>& a, const Vector<N, T>& b, NormAlgorithm<N, T> norm);
 }
 
 namespace linalg{
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T product(const Vector<N, T>& a, const Vector<N, T>& b){
         T result = 0;
         for(auto i = 0; i != N; ++i){
@@ -41,7 +41,7 @@ namespace linalg{
         return result;
     }
 
-    template <size_t P, size_t N, typename T>
+    template <int P, int N, typename T>
     T norm(const Vector<N, T>& vector){
         static_assert(P >= 1, "P cannot be less than one");
         T sum = 0;
@@ -51,17 +51,17 @@ namespace linalg{
         return pow(sum, (double)1 / P);
     }
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T euclidian(const Vector<N, T>& vector){
         return norm<2>(vector);
     }
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T manhattan(const Vector<N, T>& vector){
         return norm<1>(vector);
     }
 
-    template <size_t N, typename T>
+    template <int N, typename T>
     T distance(Vector<N, T>& a, const Vector<N, T>& b, NormAlgorithm<N, T> norm){
         const Vector<N, T> sub = a - b;
         return norm(sub);
